fix(treesitter): count what the worker emits, truncate in the parent, widen the bound
Follow-up to 005d235, from a review pass over the diff.

1. The worker bounded on len(e.decls) -- the number of declaration RANGES --
   while the parent bounds on the number of KindDecl FRAMES it receives. emit
   flattens e.decls into one record per (range, name start) pair, so a range
   declaring several names (`var a, b = ...`) makes the worker send more frames
   than its own count admitted and the parent would fail a healthy unit. The
   worker counts emitted records now (extraction.declCount).
2. The parent's three checks returned an error, which fails the unit. A bound
   the operator set is a short answer, not a protocol fault: they drop the
   frame, set extraction.overflow, and fold it into done.Truncated, which
   provider.go already turns into a partial structure capability. This is the
   truncate-and-flag shape the C3 patch asked for.
3. wire.Request.MaxRecordsPerFile was uint32: a user-set 5000000000 wrapped to
   705032704 and was applied, silently, as a bound smaller than the one asked
   for. It is uint64, which is what both comparisons already used.

Also strengthens internal/graph/rollup_bounds_test.go: its own fixture instance
gives one node twenty containers, and the comment now records exactly what each
mutation proves and what stays read-verified.

+++ internal/provider/treesitter/provider.go
@@ -236,7 +236,7 @@ func (p *Provider) IndexUnit(ctx context.Context, req provider.UnitRequest, sink
 		return finish(model.CapabilityUnavailable, model.CodeProviderUnavailable)
 	}
 	ex, err := p.parse(ctx, wire.Request{Language: l.Name, Path: fv.Path, SourceBytes: uint32(len(src)),
-		MaxRecordsPerFile: uint32(p.opts.MaxRecordsPerFile.Value())}, src)
+		MaxRecordsPerFile: uint64(p.opts.MaxRecordsPerFile.Value())}, src)
 	if err != nil {
 		return model.ProviderResult{}, err
 	}
@@ -363,7 +363,7 @@ func (p *Provider) ParseProbe(ctx context.Context, relPath string, src []byte) (
 		return Probe{}, &model.Error{Code: model.CodeResourceLimit, Message: "the file exceeds max parse file bytes"}
 	}
 	ex, err := p.parse(ctx, wire.Request{Language: l.Name, Path: relPath, SourceBytes: uint32(len(src)),
-		MaxRecordsPerFile: uint32(p.opts.MaxRecordsPerFile.Value())}, src)
+		MaxRecordsPerFile: uint64(p.opts.MaxRecordsPerFile.Value())}, src)
 	if err != nil {
 		return Probe{}, err
 	}
